Resolve an error that caused database test failure.
For some unknown reason (hope this doesn't bite me in the ass), the
in-memory database was set to share the cache. This meant every database
test was attempting to use the same database for each test, causing
errors when copying old, non-existent data between tables. Hopefully
removing the shared cache will fix the issue.

This also shows a major flaw: copying old data doesn't check that the
data exists! So that needs to be fixed. Immediately.